backend/pyproject.toml
[project]
name = "anomaly-engine"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.4",
    "paho-mqtt>=2.1.0",
    "scikit-learn>=1.8.0",
]
